+++ src/Haskellorls/Quote/Type.hs
@@ -1,0 +1,12 @@
+module Haskellorls.Quote.Type
+  ( QuoteStyle (..),
+  )
+where
+
+data QuoteStyle
+  = NoQuote
+  | SpacePadding
+  | SingleQuote
+  | DoubleQuote
+  | DynamicQuote
+  | DynamicQuoteForLink